requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.47
pandas==1.5.3
scikit-learn==1.1.3
python-dotenv==1.0.0
